Fix: Key to iterate on (Prefix only constraint) - Memtable
In my previous commit I made a huge mistake where I overlooked that we need to have a key to iterate on. My previous implementation was only finding the next after the given prefix, now we look for the next after a given key, with the prefix as a constraint.