Fix: PrefixIterate include prefix
Fixed the Prefix iterate to include the prefix itself with a smart loophole - for the starting key (for iterator creation), I pass the biggest lexicographically smaller string than the prefix. That way we include the prefix as well if it exists, since I overlooked that part in my backend logic. But this works now :)
Also there is a TODO left, that we'll need to fix (added it in Home.jsx)